minor: Rename ExampleX to SampleX in IT
