+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule4852classannotations/sample2/package-info.java
@@ -1,0 +1,5 @@
+/** This is a package. */
+@Deprecated
+// testing
+@SuppressWarnings("...")
+package com.google.checkstyle.test.chapter4formatting.rule4852classannotations.sample2;
