+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule4852classannotations/sample3/package-info.java
@@ -4,7 +4,7 @@
 */
 
 @ParentPackage("person") // violation below 'Annotation 'Name' should be alone on line'
-@Name("/") package com.google.checkstyle.test.chapter4formatting.rule4852classannotations.example3;
+@Name("/") package com.google.checkstyle.test.chapter4formatting.rule4852classannotations.sample3;
 
 import com.google.checkstyle.test.chapter4formatting.rule4852classannotations.Name;
 import com.google.checkstyle.test.chapter4formatting.rule4852classannotations.ParentPackage;
