+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule4852classannotations/sample5/package-info.java
@@ -5,7 +5,7 @@
 @ParentPackage("val")
 @Name("1")
 @SizePackage("2")
-package com.google.checkstyle.test.chapter4formatting.rule4852classannotations.example5;
+package com.google.checkstyle.test.chapter4formatting.rule4852classannotations.sample5;
 
 import com.google.checkstyle.test.chapter4formatting.rule4852classannotations.Name;
 import com.google.checkstyle.test.chapter4formatting.rule4852classannotations.ParentPackage;
